Return actual error message from the server instead of hardcoded one.
This commit adds a new `GetErr()` method that takes the server response
and returns actual error message from the server in case of an `ERR`
response. The `IsOk()` checks in the client initiation and`Cmd()`
are replaced with this new method to get actual error messages.

There are no breaking changes.